move default layout component to layouts directory.
